#216: upload images

+++ src/main/scala/no/ndla/imageapi/model/api/Image.scala
@@ -7,5 +7,5 @@ import scala.annotation.meta.field
 
 @ApiModel(description = "Url and size information about the image")
 case class Image(@(ApiModelProperty@field)(description = "The full url to where the image can be downloaded") url: String,
-                 @(ApiModelProperty@field)(description = "The size of the image in bytes") size: Int,
+                 @(ApiModelProperty@field)(description = "The size of the image in bytes") size: Long,
                  @(ApiModelProperty@field)(description = "The mimetype of the image") contentType: String)
